Update worker to decouple relevant parts of compiler-cli
compiler-cli that is configurable will be only added to the worker
as dependency when the Angular worker is requested. This allows us to
actually compile `packages/compiler-cli` with the vanilla TS worker.

We are still using some types and foundational logic from compiler-cli
for the virtual FS etc; but those parts will originate from the compiler
CLI version from npm in vanilla compilations. In Angular compilations,
the compiler will transition to use the configured version even for the
foundational logic, as well as for any compilation processing of Angular
code.