Fixes to support compiler-cli from HEAD (#38)
When trying out using the compiler-cli from HEAD, we couldn't do it
before this commit because there was no `JsInfo`. This commit fixes
this.

In addition, the symlinks created, as noted, only worked for `.runfiles`
runtime execution— so this resulted in some build failures when
compiling the worker `ts_project`.

We fix this by using `target_file` symlinks, that are automatically
adjusted by Bazel based on where the symlink is needed. i.e. in runfiles
dir, or in `bazel-bin`. This solves this nicely.

Unfortunately though, the final problem is that at runtime, due to the
new `target_file` option, the symlinks are no longer relative pointing
to the pnpm root inside `.runfiles`, but instead point to the execroot.
That is because they are now "absolute". Issues now arise when NodeJS
tries to find transitive dependencies of e.g. `@angular/compiler-cli`.
Such issues arise because `rules_js` generally doesn't support escaping
the `.runfiles` to another "safe" root. In this case that is actually
desirable, so we turn that patching off to support jumping between the
roots (and enable transitive dependency resolution).